+++ src/resolvers/Mutation/registerVideo/prisma.test.ts
@@ -23,9 +23,9 @@ import { afterAll, beforeAll, beforeEach, describe, expect, test } from "vitest"
 import { cleanPrisma } from "../../../test/cleanPrisma.js";
 import { Ok } from "../../../utils/Result.js";
 import { ResolverDeps } from "../../index.js";
-import { register } from "./registerVideo.js";
+import { register } from "./prisma.js";
 
-describe("Mutation.registerVideo", () => {
+describe("Register video by Prisma", () => {
   let prisma: ResolverDeps["prisma"];
 
   beforeAll(async () => {
